refactor(tag): move TagRepository classes into tag package
