+++ tracker-api/src/test/java/com/sloyardms/trackerapi/controller/TagControllerIntegrationTests.java
@@ -4,7 +4,7 @@ import com.sloyardms.trackerapi.tag.dto.TagCreateDto;
 import com.sloyardms.trackerapi.tag.dto.TagDto;
 import com.sloyardms.trackerapi.tag.dto.TagUpdateDto;
 import com.sloyardms.trackerapi.user.dto.UserCreateDto;
-import com.sloyardms.trackerapi.repository.TagRepository;
+import com.sloyardms.trackerapi.tag.TagRepository;
 import com.sloyardms.trackerapi.security.DevSecurityConfig;
 import com.sloyardms.trackerapi.security.FakeAuthFilter;
 import io.restassured.RestAssured;
